Nest admin-only API types
