+++ packages/ui-extensions/src/api.ts
@@ -74,24 +74,3 @@ export interface I18n {
    */
   translate: I18nTranslate;
 }
-
-export interface Intents {
-  /**
-   * The URL that was used to launch the intent.
-   */
-  launchUrl?: string | URL;
-}
-
-export interface Navigation {
-  /**
-   * A method to navigate to a specific route.
-   */
-  navigate: (url: string | URL) => void;
-}
-
-export interface Data {
-  /**
-   * Information about the currently viewed or selected items.
-   */
-  selected: {id: string}[];
-}
