+++ packages/ui-extensions/src/surfaces/admin/api.ts
@@ -1,5 +1,5 @@
 export type {I18n, I18nTranslate} from '../../api';
-export type {StandardApi} from './api/standard/standard';
+export type {StandardApi, Navigation, Intents} from './api/standard/standard';
 export type {CustomerSegmentationTemplateApi} from './api/customer-segmentation-template/customer-segmentation-template';
 export type {ActionExtensionApi} from './api/action/action';
 export type {BlockExtensionApi} from './api/block/block';
