+++ packages/ui-extensions/src/surfaces/admin/api/shared.ts
@@ -1,0 +1,6 @@
+export interface Data {
+  /**
+   * Information about the currently viewed or selected items.
+   */
+  selected: {id: string}[];
+}
